*: update minimal dependencies for gccgo support
Also:
 - Update golang/dep
 - Enable pruning (removes ~6M LOC from vendor)
 - Move all possibile dependencies to tagged releases

This commit enables building on gccgo 8.0 (pre-release), though linker
flags will require additional massaging, see
https://groups.google.com/forum/#!msg/golang-nuts/SlGCPYkjxo4/4DjcjXRCqAkJ

$ dep-diff-parser
https://github.com/apache/thrift/compare/a533ea16...327ebb6c
https://github.com/axiomhq/hyperloglog/compare/83009472...baba800b
https://github.com/dgryski/go-metro/compare/0f647357...280f6062
https://github.com/go-ole/go-ole/compare/2ecd927e...a41e3c4b
https://github.com/golang/dep/compare/8ddfc8af...37d9ea0a
https://github.com/golang/protobuf/compare/1643683e...92554152
https://github.com/google/go-github/compare/996760c5...e48060a2
https://github.com/grpc-ecosystem/grpc-gateway/compare/8879672d...07f5e797
https://github.com/jmank88/nuts/compare/a1e02c78...8b28145d
https://github.com/openzipkin/zipkin-go-opentracing/compare/9c88fa03...45e90b00
https://github.com/pelletier/go-toml/compare/4e9e0ee1...acdc4509
https://github.com/petermattis/goid/compare/3db12ebb...176e84a9
https://github.com/sasha-s/go-deadlock/compare/565eb443...03d40e5d
https://github.com/satori/go.uuid/compare/5bf94b69...f58768cc
https://github.com/dominikh/go-tools/compare/80dfe556...d73ab98e

Release note: None

+++ pkg/cmd/github-post/main_test.go
@@ -150,7 +150,7 @@ Stress build found a failed test: %s`,
 					if length := len(*issue.Body); length > githubIssueBodyMaximumLength {
 						t.Fatalf("issue length %d exceeds (undocumented) maximum %d", length, githubIssueBodyMaximumLength)
 					}
-					return &github.Issue{ID: github.Int(issueID)}, nil, nil
+					return &github.Issue{ID: github.Int64(issueID)}, nil, nil
 				}
 				searchIssues := func(_ context.Context, query string, opt *github.SearchOptions) (*github.IssuesSearchResult, *github.Response, error) {
 					total := 0
